--- .eslintrc.yml
@@ -1,17 +1,0 @@
-env:
-  browser: true
-  es2021: true
-extends:
-  - eslint:recommended
-  - plugin:react/recommended
-  - plugin:@typescript-eslint/recommended
-parser: '@typescript-eslint/parser'
-parserOptions:
-  ecmaFeatures:
-    jsx: true
-  ecmaVersion: latest
-  sourceType: module
-plugins:
-  - react
-  - '@typescript-eslint'
-rules: { 'react/jsx-uses-react': 'off', 'react/react-in-jsx-scope': 'off' }
