--- .prettierrc.yml
@@ -1,6 +1,0 @@
-# .prettierrc or .prettierrc.yaml
-trailingComma: 'es5'
-tabWidth: 2
-semi: true
-singleQuote: true
-printWidth: 80
