+++ entrypoints/options/components/import_button.tsx
@@ -1,3 +1,5 @@
+import React from 'react';
+
 const onClick = async (
   event: React.MouseEvent<HTMLButtonElement, MouseEvent>
 ) => {
