+++ entrypoints/options/index.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html lang="ja">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Pixiv Filter 設定ページ</title>
+    <meta name="manifest.open_in_tab" content="true" />
+    <!-- <meta name="manifest.chrome_style" content="true" /> -->
+    <!-- <meta name="manifest.browser_style" content="true" /> -->
+    <!-- Set include/exclude if the page should be removed from some builds -->
+    <!-- <meta name="manifest.include" content="['chrome', ...]" />
+    <meta name="manifest.exclude" content="['chrome', ...]" /> -->
+  </head>
+  <body>
+    <div id="root"></div>
+    <script type="module" src="./main.tsx"></script>
+  </body>
+</html>
