+++ entrypoints/options/option_page.css
@@ -1,0 +1,49 @@
+* {
+  font-size: 2vh;
+}
+
+.App {
+  margin: 0;
+  padding: 0;
+  height: 90svh;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.title-container {
+  display: flex;
+  flex-direction: row;
+}
+
+.ng-config-container {
+  height: 50%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: start;
+}
+
+select {
+  height: 100%;
+  /* min-height: 100%; */
+  /* max-height: 100%; */
+  /* max-width: 10svw; */
+  min-width: 30svw;
+}
+
+.select-container {
+  height: 100%;
+  display: flex;
+  flex-direction: row;
+  align-items: end;
+  gap: 0.5rem;
+  padding-bottom: 0.5rem;
+}
+
+.import-export-buttons {
+  display: flex;
+  flex-direction: row;
+  gap: 0.5rem;
+}
