+++ entrypoints/types/type.d.ts
