+++ entrypoints/works_style.ts
@@ -1,3 +1,5 @@
+export default defineUnlistedScript(() => {});
+
 export const WorksStyle = {
   hidden: 'li { visibility: hidden; }',
   visible: 'li { visibility: visible; }',
