+++ package.json
@@ -1,34 +1,29 @@
 {
-  "name": "pixiv-filter",
+  "name": "pixiv-filter-for-wxt",
+  "description": "PixivでNG登録したユーザーやタグの作品を非表示にします。",
   "private": true,
-  "version": "0.0.0",
+  "version": "1.3.1",
   "type": "module",
   "scripts": {
-    "dev": "vite",
-    "build": "tsc && vite build",
-    "preview": "vite preview"
+    "dev": "wxt",
+    "dev:firefox": "wxt -b firefox",
+    "build": "wxt build",
+    "build:firefox": "wxt build -b firefox",
+    "zip": "wxt zip",
+    "zip:firefox": "wxt zip -b firefox",
+    "compile": "tsc --noEmit",
+    "postinstall": "wxt prepare"
   },
   "dependencies": {
-    "@types/chrome": "^0.0.259",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "tippy.js": "^6.3.7"
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1"
   },
   "devDependencies": {
-    "@crxjs/vite-plugin": "^2.0.0-beta.23",
-    "@types/react": "^18.2.48",
-    "@types/react-dom": "^18.2.18",
-    "@types/wicg-file-system-access": "^2023.10.4",
-    "@typescript-eslint/eslint-plugin": "^6.19.1",
-    "@typescript-eslint/parser": "^6.19.1",
-    "@vitejs/plugin-react": "2.2.0",
-    "eslint": "^8.56.0",
-    "eslint-config-airbnb-typescript": "^17.1.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-react": "^7.33.2",
-    "prettier": "3.2.4",
-    "typescript": "^5.3.3",
-    "vite": "3.2.8"
+    "@types/chrome": "^0.0.278",
+    "@types/react": "^18.3.11",
+    "@types/react-dom": "^18.3.1",
+    "@wxt-dev/module-react": "^1.1.1",
+    "typescript": "^5.6.3",
+    "wxt": "^0.19.11"
   }
 }
